Remove remaining references to "relay"
The old "relay" name still existed in the code. For consistency these
have all been changes to "spout".

COMPATIBILITY NOTE: This changes the metrics names published to the
NATS monitor subject.
